modified dialog plugin so it scrolls etc. +
modified to take optional parameters that included locationX and locationY for placement.  Modified setText and related functions to place text in the correct place and scroll it. Also modified so setText will automatically display the window if it is currently not visible.
I have not attempted to check animation of text, and the plugin could use more modification to allow custom width as well.  Currently I put in the optional parameter for it but its value is not filtered down through the code.